split PS constants out of models
