+++ apps/trackers/tests/test_file_offer.py
@@ -4,8 +4,8 @@ Test cases for tracker suggestion generation
 
 import pytest
 
-from osidb.dmodels import PsUpdateStream
-from osidb.models import Affect, Flaw, Impact, UbiPackage
+from osidb.dmodels import PsUpdateStream, UbiPackage
+from osidb.models import Affect, Flaw, Impact
 from osidb.tests.factories import (
     AffectFactory,
     FlawFactory,
